迎新系统整合, 确定目录结构
